chore: remove jest-config dependency from jest-runneer
