+++ packages/jest-config/src/utils.ts
@@ -117,134 +117,9 @@ export const _replaceRootDirTags = <T extends ReplaceRootDirConfigValues>(
   return config;
 };
 
-export const resolveWithPrefix = (
-  resolver: string | undefined | null,
-  {
-    filePath,
-    humanOptionName,
-    optionName,
-    prefix,
-    rootDir,
-  }: {
-    filePath: string;
-    humanOptionName: string;
-    optionName: string;
-    prefix: string;
-    rootDir: Config.Path;
-  },
-): string => {
-  const fileName = replaceRootDirInPath(rootDir, filePath);
-  let module = Resolver.findNodeModule(`${prefix}${fileName}`, {
-    basedir: rootDir,
-    resolver: resolver || undefined,
-  });
-  if (module) {
-    return module;
-  }
-
-  try {
-    return require.resolve(`${prefix}${fileName}`);
-  } catch {}
-
-  module = Resolver.findNodeModule(fileName, {
-    basedir: rootDir,
-    resolver: resolver || undefined,
-  });
-  if (module) {
-    return module;
-  }
-
-  try {
-    return require.resolve(fileName);
-  } catch {}
-
-  throw createValidationError(
-    `  ${humanOptionName} ${chalk.bold(
-      fileName,
-    )} cannot be found. Make sure the ${chalk.bold(
-      optionName,
-    )} configuration option points to an existing node module.`,
-  );
-};
-
-/**
- * Finds the test environment to use:
- *
- * 1. looks for jest-environment-<name> relative to project.
- * 1. looks for jest-environment-<name> relative to Jest.
- * 1. looks for <name> relative to project.
- * 1. looks for <name> relative to Jest.
- */
-export const getTestEnvironment = ({
-  rootDir,
-  testEnvironment: filePath,
-}: {
-  rootDir: Config.Path;
-  testEnvironment: string;
-}): string =>
-  resolveWithPrefix(undefined, {
-    filePath,
-    humanOptionName: 'Test environment',
-    optionName: 'testEnvironment',
-    prefix: 'jest-environment-',
-    rootDir,
-  });
-
-/**
- * Finds the watch plugins to use:
- *
- * 1. looks for jest-watch-<name> relative to project.
- * 1. looks for jest-watch-<name> relative to Jest.
- * 1. looks for <name> relative to project.
- * 1. looks for <name> relative to Jest.
- */
-export const getWatchPlugin = (
-  resolver: string | undefined | null,
-  {filePath, rootDir}: {filePath: string; rootDir: Config.Path},
-): string =>
-  resolveWithPrefix(resolver, {
-    filePath,
-    humanOptionName: 'Watch plugin',
-    optionName: 'watchPlugins',
-    prefix: 'jest-watch-',
-    rootDir,
-  });
-
-/**
- * Finds the runner to use:
- *
- * 1. looks for jest-runner-<name> relative to project.
- * 1. looks for jest-runner-<name> relative to Jest.
- * 1. looks for <name> relative to project.
- * 1. looks for <name> relative to Jest.
- */
-export const getRunner = (
-  resolver: string | undefined | null,
-  {filePath, rootDir}: {filePath: string; rootDir: Config.Path},
-): string =>
-  resolveWithPrefix(resolver, {
-    filePath,
-    humanOptionName: 'Jest Runner',
-    optionName: 'runner',
-    prefix: 'jest-runner-',
-    rootDir,
-  });
-
 type JSONString = string & {readonly $$type: never}; // newtype
 export const isJSONString = (text?: JSONString | string): text is JSONString =>
   text != null &&
   typeof text === 'string' &&
   text.startsWith('{') &&
   text.endsWith('}');
-
-export const getSequencer = (
-  resolver: string | undefined | null,
-  {filePath, rootDir}: {filePath: string; rootDir: Config.Path},
-): string =>
-  resolveWithPrefix(resolver, {
-    filePath,
-    humanOptionName: 'Jest Sequencer',
-    optionName: 'testSequencer',
-    prefix: 'jest-sequencer-',
-    rootDir,
-  });
