+++ packages/jest-resolve/tsconfig.json
@@ -7,6 +7,7 @@
   "references": [
     {"path": "../jest-haste-map"},
     {"path": "../jest-types"},
-    {"path": "../jest-util"}
+    {"path": "../jest-util"},
+    {"path": "../jest-validate"}
   ]
 }
