+++ packages/jest-runner/tsconfig.json
@@ -5,7 +5,6 @@
     "outDir": "build"
   },
   "references": [
-    {"path": "../jest-config"},
     {"path": "../jest-console"},
     {"path": "../jest-docblock"},
     {"path": "../jest-environment"},
